some more

+++ aloeda-service/install.sh
@@ -4,14 +4,18 @@
 # Description : Simple app service installer
 
 # vars
-BUILD_CONTAINERS=false     # build containers
-RUN_CONTAINERS=false       # run containers on ready
+COMPILE_CONTAINERS=true     # build containers
+RUN_CONTAINERS=true         # run containers on ready
+STOP_ALL=true               # reset all images and containers
+PRUNE_ALL=true              # prune all images and containers
 
-while getopts b:r: option
+while getopts c:r:a:p: option
 do
     case "${option}"
     in
-    b) BUILD_CONTAINERS=${OPTARG};;
+    a) PRUNE_ALL=${OPTARG};;
+    a) STOP_ALL=${OPTARG};;
+    c) COMPILE_CONTAINERS=${OPTARG};;
     r) RUN_CONTAINERS=${OPTARG};;
     esac
 done
@@ -19,13 +23,20 @@ done
 # build
 ./gradlew clean assemble
 
-# cleanup
-docker stop $(docker ps -a -q) && docker rm -f $(docker ps -a -q)
-docker image prune -f
-docker volume prune -f
-docker network prune -f
+# stop all
+if [[ "$STOP_ALL" == 'true' ]]; then
+    docker stop $(docker ps -a -q) && docker rm -f $(docker ps -a -q)
+fi
 
-if [[ "$BUILD_CONTAINERS" == 'true' ]]; then
+# prune all
+if [[ "$PRUNE_ALL" == 'true' ]]; then
+    docker image prune -f
+    docker volume prune -f
+    docker network prune -f
+fi
+
+# prep all
+if [[ "$COMPILE_CONTAINERS" == 'true' ]]; then
     # switch to installer
     cd installer
 
@@ -34,6 +45,7 @@ if [[ "$BUILD_CONTAINERS" == 'true' ]]; then
     docker-compose pull
 fi
 
+# run all
 if [[ "$RUN_CONTAINERS" == 'true' ]]; then
-    docker-compose up --build -d --scale app_redis_replica=2
+    docker-compose up --build -d --scale app_redis_replica=3
 fi
